ext_sc/requirements.txt
apturl==0.5.2
asgiref==3.4.1
attrs==21.4.0
autokey==0.95.10
Automat==20.2.0
autopep8==1.6.0
backports.entry-points-selectable==1.1.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.3
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.2.11
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.3.1
future==0.18.2
gpg===1.13.1-unknown
gunicorn==20.1.0
httplib2==0.14.0
hyperlink==21.0.0
idna==2.8
incremental==21.3.0
itemadapter==0.6.0
itemloaders==1.0.4
jmespath==1.0.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
parsel==1.6.0
pexpect==4.6.0
Pillow==7.0.0
platformdirs==2.4.0
Protego==0.2.1
protobuf==3.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.8.0
pycups==1.9.73
PyDispatcher==2.0.5
PyGObject==3.36.0
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.3.0
pyOpenSSL==22.0.0
PyQt5==5.14.1
pyRFC3339==1.1
PySocks==1.6.8
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-decouple==3.5
python-xlib==0.23
pytz==2021.3
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
queuelib==1.6.2
reportlab==3.5.34
requests==2.22.0
requests-file==1.5.1
requests-unixsocket==0.2.0
Scrapy==2.6.1
SecretStorage==2.3.1
service-identity==21.1.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sqlparse==0.4.2
systemd-python==234
tldextract==3.3.0
toml==0.10.2
torbrowser-launcher==0.3.2
Twisted==22.4.0
typing-extensions==4.2.0
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
Unipath==1.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.9.0
w3lib==1.22.0
wadllib==1.3.3
whitenoise==5.3.0
xkit==0.0.0
youtube-dl==2020.3.24
zope.interface==5.4.0
